requirements.txt
# ============================================
# YouTube Knowledge Assistant - Dependencies
# ============================================
# Install: pip install -r requirements.txt

# --- Core Framework ---
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.2.0

# --- LLM Providers ---
openai>=1.50.0

# --- Vector Store ---
chromadb>=0.5.0

# --- Embeddings ---
tiktoken>=0.7.0
sentence-transformers>=3.0.0    # For HuggingFace embeddings fallback

# --- YouTube Data ---
youtube-transcript-api>=0.6.2
yt-dlp>=2024.10.0               # Playlist/channel resolution
google-api-python-client>=2.140.0  # YouTube Data API v3

# --- Audio Transcription (Whisper fallback) ---
openai-whisper>=20231117

# --- Sentiment Analysis ---
vaderSentiment>=3.3.2
pandas>=2.2.0

# --- Frontend ---
streamlit>=1.38.0
streamlit-chat>=0.1.1

# --- Utilities ---
python-dotenv>=1.0.1
pydantic>=2.9.0

# --- Testing ---
pytest>=8.3.0
